pyproject.toml
[project]
name = "app-vector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "async>=0.6.2",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "jose>=1.0.0",
    "jwt>=1.3.1",
    "passlib>=1.7.4",
    "pg>=0.1",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.36",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.4",
    "taskipy>=1.14.1",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark test as async"
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 79
extend-exclude = ['migrations']

target-version = "py312"

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'




[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
dev = 'fastapi dev app/main.py'
run = 'fastapi run app/main.py'
test = 'pytest -s -x --cov=app -vv'
test_cov = 'coverage html'

